Revert blog layout
The website was in the process of becoming a blog layout, however due to a lack of time and effort, I've now decided against having my website as a blog. Instead it is more in a CV format. As such, this PR is to revert back to the CV layout.